make groups modal and role modal not depending on eachother and Change parameters in function editGroups and editRoles to (Models) change @ViewChild('inputValue') inputValue: ElementRef;
