+++ src/app/core/main/components/role-setting-modal/role-setting-modal.component.css
@@ -1,0 +1,3 @@
+.p-button.p-button-icon-only {
+  padding: 0.5rem 1.3rem;
+}
