+++ src/app/core/main/components/role-setting-modal/setting-User-Dto.ts
@@ -1,0 +1,6 @@
+export interface LoginUserDataDTO {
+    userId: string,
+    name: string,
+    groups: string[],
+    role : string,
+}
